refactor: Split `neighborlist` module into submodules
